+++ cfw/nextui/nextui.go
@@ -11,8 +11,8 @@ import (
 var embeddedFiles embed.FS
 
 var (
-	Platforms       = jsonutil.MustLoadJSONMap[string, []string](embeddedFiles, "data/platforms.json", "cfw/nextui")
-	SaveDirectories = jsonutil.MustLoadJSONMap[string, []string](embeddedFiles, "data/save_directories.json", "cfw/nextui")
+	Platforms       = jsonutil.MustLoadJSONMap[string, []string](embeddedFiles, "data/platforms.json")
+	SaveDirectories = jsonutil.MustLoadJSONMap[string, []string](embeddedFiles, "data/save_directories.json")
 )
 
 func GetBasePath() string {